feat(template): transparent fills/borders on table cells (#76)
- Widen CellStyle.backgroundColor and borderColor to string | null;
  null is the sentinel for "transparent" (no fill / no stroke),
  matching the existing PageDefinition.backgroundColor convention.
- Gate the PDFKit renderer's cell border .stroke() on a non-null
  borderColor; existing falsy guards already cover the fill side.
- Add a NullableColorInput control (checker swatch + Clear/Color
  toggle, themed border) and wire it for header background, row
  background, and cell border colour in the right panel.
- Split LoopFieldProps.tsx into TableColumnsSection.tsx and
  TableStyleSections.tsx so each file stays under the 300-line cap.
- Canvas: merge per-column style over rowStyle in tableBodyRows so
  per-column align (and other overrides) actually take effect on the
  Fabric preview, matching core/render/loop.ts. Switch body cell
  textbox to a left-anchored origin so textAlign reliably governs
  text placement within the column.
- Rename right-panel labels to disambiguate header vs row text vs
  background vs border colour fields.
- Tests: round-trip null backgroundColor/borderColor through save +
  readManifest; render with transparent header bg and row bg without
  error; PDF byte length is smaller when fills/strokes are skipped.
